Runs as applet - KeyListener is private in Applet class, hardDrop() added. Works w/o visual quirks.
