+++ .gitignore
@@ -2,3 +2,6 @@
 /Tetris.class
 /TetrisKeyListener.class
 /Tetris$1.class
+/TetrisGame$1.class
+/TetrisGame$TetrisActionListener.class
+/Tetris$TetrisKeyListener.class
